feat: :bug: Fix Rerender NavBar, navigate Artist Profile
The ArtistCard and SongItem components have been updated to use Link to navigate to the artist profile and song page, respectively. The SongFieldData component has been updated to use the correct path for the Link to the artist profile. The Search component has been updated to remove the StickyNavbar component.