style: :art: Display homepage for user and guest. Format Audio Player. (Full Responsive)
Update frontend components and add new route for uploading song  The diff includes updates to the frontend components such as NavBar, PlaylistCard, SongProfile, and the addition of a new route for uploading a song. The new route is protected by a LoginedRoute component, which ensures that only logged-in users can access it. The UploadSong component has also been updated to include a new image input field and a new step for selecting artists. Additionally, the Homepage component has been moved to a new directory and updated to include a new layout.