refactor: :lipstick: Update For PlayList UI/UX
Backend:  1. Added methods to add and remove songs from a song list in the SongListController. 2. Added a use of the DB facade to retrieve song lists in the SongListController.  Frontend:  1. Added routes for adding and removing songs from a song list in api.php. 2. Removed DeleteConfirmModal, PlayListItem, Playlists, SearchField, and UpdatePlayListModal components. 3. Updated the import paths for some components in Library/index.jsx. 4. Updated the import paths for some components in router.jsx. 5. Added an alias for the '~' import path in vite.config.js.